refactor:  컴포넌트 구조 재정리
- Header/Footer를 widgets 레이어로 이동
- components 폴더 정리 및 불필요한 타입 제거
