+++ src/App.tsx
@@ -1,20 +1,20 @@
-import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./components/Header.tsx"
-import Footer from "./components/Footer.tsx"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import { BrowserRouter as Router } from 'react-router-dom';
+import { Header } from './widgets/Header';
+import { Footer } from './widgets/Footer';
+import PostsManagerPage from './pages/PostsManagerPage.tsx';
 
 const App = () => {
   return (
     <Router>
-      <div className="flex flex-col min-h-screen">
+      <div className='flex flex-col min-h-screen'>
         <Header />
-        <main className="flex-grow container mx-auto px-4 py-8">
+        <main className='flex-grow container mx-auto px-4 py-8'>
           <PostsManagerPage />
         </main>
         <Footer />
       </div>
     </Router>
-  )
-}
+  );
+};
 
-export default App
+export default App;
