--- src/components/index.tsx
@@ -1,20 +1,0 @@
-export { Button } from '../shared/ui/Button';
-export { Input } from '../shared/ui/Input';
-export { Card, CardHeader, CardTitle, CardContent } from '../shared/ui/Card';
-export { Textarea } from '../shared/ui/Textarea';
-export {
-  Select,
-  SelectGroup,
-  SelectValue,
-  SelectTrigger,
-  SelectContent,
-  SelectItem,
-} from '../shared/ui/Select';
-export {
-  Dialog,
-  DialogTrigger,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from '../shared/ui/Dialog';
-export { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from '../shared/ui/Table';
