+++ src/widgets/Footer/Footer.tsx
